Backend/requirements.txt
fastapi
pillow
roboflow
transformers
torch
verovio
torchvision
uvicorn
tiktoken
accelerate
python-multipart